calc blowout now method for game
